+++ crates/frm-bin/src/common/env_vars.rs
@@ -1,0 +1,11 @@
+// Copyright (c) 2025-2026 Michael S. Klishin and Contributors
+//
+// Licensed under the Apache License, Version 2.0 <LICENSE-APACHE or
+// https://www.apache.org/licenses/LICENSE-2.0> or the MIT license
+// <LICENSE-MIT or https://opensource.org/licenses/MIT>, at your
+// option. This file may not be copied, modified, or distributed
+// except according to those terms.
+
+pub const RABBITMQ_HOME: &str = "RABBITMQ_HOME";
+pub const FRM_DIR: &str = "FRM_DIR";
+pub const FRM_SHELL: &str = "FRM_SHELL";