+++ crates/frm-bin/src/common/http.rs
@@ -1,0 +1,9 @@
+// Copyright (c) 2025-2026 Michael S. Klishin and Contributors
+//
+// Licensed under the Apache License, Version 2.0 <LICENSE-APACHE or
+// https://www.apache.org/licenses/LICENSE-2.0> or the MIT license
+// <LICENSE-MIT or https://opensource.org/licenses/MIT>, at your
+// option. This file may not be copied, modified, or distributed
+// except according to those terms.
+
+pub const USER_AGENT: &str = concat!("frm/", env!("CARGO_PKG_VERSION"));