+++ crates/frm-bin/src/common/mod.rs
@@ -1,0 +1,12 @@
+// Copyright (c) 2025-2026 Michael S. Klishin and Contributors
+//
+// Licensed under the Apache License, Version 2.0 <LICENSE-APACHE or
+// https://www.apache.org/licenses/LICENSE-2.0> or the MIT license
+// <LICENSE-MIT or https://opensource.org/licenses/MIT>, at your
+// option. This file may not be copied, modified, or distributed
+// except according to those terms.
+
+pub mod cli_tools;
+pub mod env_vars;
+pub mod http;
+pub mod urls;