+++ crates/frm-bin/src/common/urls.rs
@@ -1,0 +1,15 @@
+// Copyright (c) 2025-2026 Michael S. Klishin and Contributors
+//
+// Licensed under the Apache License, Version 2.0 <LICENSE-APACHE or
+// https://www.apache.org/licenses/LICENSE-2.0> or the MIT license
+// <LICENSE-MIT or https://opensource.org/licenses/MIT>, at your
+// option. This file may not be copied, modified, or distributed
+// except according to those terms.
+
+pub const SERVER_PACKAGES_API_URL: &str =
+    "https://api.github.com/repos/rabbitmq/server-packages/releases";
+
+pub const RABBITMQ_SERVER_API_URL: &str =
+    "https://api.github.com/repos/rabbitmq/rabbitmq-server/releases";
+
+pub const RABBITMQ_SIGNING_KEY_URL: &str = "https://github.com/rabbitmq/signing-keys/releases/download/3.0/rabbitmq-release-signing-key.asc";